Added Lot size
calculation for added lot size, updated red to brighter red due to user complain eye pain

+++ app/src/main/java/com/example/tracer/model/Trade.java
@@ -12,39 +12,39 @@ public class Trade {
     @ColumnInfo(name = "name")
     private String cryptoName;
     @ColumnInfo(name = "buyPrice")
-    private float buyPrice;
+    private double buyPrice;
     @ColumnInfo(name = "sellPrice")
-    private float sellPrice;
+    private double sellPrice;
     @ColumnInfo(name = "amountBought")
-    private float amountBought;
+    private double amountBought;
     @ColumnInfo(name = "askPrice")
-    private float askPrice;
+    private double askPrice;
     @ColumnInfo(name = "profitLossActual")
-    private float profitLossActual;
+    private double profitLossActual;
     @ColumnInfo(name = "profitLossPercent")
-    private float profitLossPercent;
+    private double profitLossPercent;
 
-    public float getProfitLossActual() {
+    public double getProfitLossActual() {
         return profitLossActual;
     }
 
-    public void setProfitLossActual(float profitLoss) {
+    public void setProfitLossActual(double profitLoss) {
         this.profitLossActual = profitLoss;
     }
 
-    public float getProfitLossPercent() {
+    public double getProfitLossPercent() {
         return profitLossPercent;
     }
 
-    public void setProfitLossPercent(float profitLossPercent) {
+    public void setProfitLossPercent(double profitLossPercent) {
         this.profitLossPercent = profitLossPercent;
     }
 
-    public float getAskPrice() {
+    public double getAskPrice() {
         return askPrice;
     }
 
-    public void setAskPrice(float askPrice) {
+    public void setAskPrice(double askPrice) {
         this.askPrice = askPrice;
     }
 
@@ -65,31 +65,31 @@ public class Trade {
         this.cryptoName = cryptoName;
     }
 
-    public float getBuyPrice() {
+    public double getBuyPrice() {
         return buyPrice;
     }
 
-    public void setBuyPrice(float buyPrice) {
+    public void setBuyPrice(double buyPrice) {
         this.buyPrice = buyPrice;
     }
 
-    public float getSellPrice() {
+    public double getSellPrice() {
         return sellPrice;
     }
 
-    public void setSellPrice(float sellPrice) {
+    public void setSellPrice(double sellPrice) {
         this.sellPrice = sellPrice;
     }
 
-    public float getAmountBought() {
+    public double getAmountBought() {
         return amountBought;
     }
 
-    public void setAmountBought(float amountBought) {
+    public void setAmountBought(double amountBought) {
         this.amountBought = amountBought;
     }
 
-    public Trade(String cryptoName, float buyPrice, float sellPrice, float amountBought) {
+    public Trade(String cryptoName, double buyPrice, double sellPrice, double amountBought) {
         this.id = System.currentTimeMillis();
         this.cryptoName = cryptoName;
         this.buyPrice = buyPrice;
